requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
attrs==23.2.0
boto3==1.34.131
botocore==1.34.131
certifi==2024.6.2
click==8.1.7
cmake==3.29.5
colorama==0.4.6
dlib==19.24.4
face-recognition==1.3.0
face_recognition_models==0.3.0
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
joblib==1.4.2
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.4
pillow==10.3.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.0.0
six==1.16.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
yarl==1.9.4